rgw/sfs: add ability to list buckets via meta_list_*() functions
This is a nasty hack to make meta_list_keys_*() work with both
users and buckets.  What we _should_ do is create some extra class that
knows how to do meta key listing of various types of object, and pass an
instance of that around via the handle arguments to these functions.
This would allow that as-yet-nonexistent class to keep track of max and
truncated in meta_list_keys_next().  Instead, the immediate nasty hack is
to make handle point at a const string which indicates what type of
metadata we're dealing with, then check that value and call an instance
of the appropriate rgw::sal::sfs::sqlite::SQLite* class to get the
metdata we care about.

Fixes: https://github.com/aquarist-labs/s3gw/issues/64
Signed-off-by: Tim Serong <tserong@suse.com>