rgw/sfs: Get user by access_key when user has multiple keys
Added a new table in sqlite's metadata to store the access keys related
to a user.
That way, when querying by access key we query first that table to
obtain the user id.

Could not implement in the user table itself because access keys are
stored in a blob and we'd need to first decode the blob in order to
query by access key.

Removed the helper columsn in users table for access keys as it was only
supporting the first key and it's no longer used.

Fixes: https://github.com/aquarist-labs/s3gw/issues/83
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>