rgw/sfs: Detect metadata compatibility issues.
Adds a check on metadata compatibility when the `SFStore` initialises.

It creates a temporary copy of the metadata database and tries to sync
the schema on it.
If there are issues (exeptions) or any table is dropped due to
incompatibilities between the data base and the actual metadata schema
it throws a `sqlite_sync_exception`.

The execption is catched when the `SFStore` is instantiated and the
error is logged before exiting `rgw`.

Previous database is not altered as all checks are done in the temporary
database so we don't have data loss.

Fixes: https://github.com/aquarist-labs/s3gw/issues/178
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>