rgw/sfs: Generate delete markers for non existing objects
Adds code for adding a delete marker when deleting an object that
doesn't exist.

According to the AWS documentation https://docs.aws.amazon.com/cli/latest/reference/s3api/delete-object.html

when deleteing an object it should reply no errors even if the object
doesn't exist.

This means we create a delete-marker version for the object when
versioning is enabled.

This also fixes the s3-test
test_versioning_multi_object_delete_with_marker_create which is exactly
that test case.

When undeleting delete markers (or deleting all versions of an object),
if the object ends up having no version we just remove it from the
in-memory objects map.

The Garbage collector could then permanently remove those objects
without a version which are, in fact, no longer reachable.

Fixes: https://github.com/aquarist-labs/s3gw/issues/186
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>