common/perf_counters: Add index to data struct
Add and carry around the perf counter index. This makes the index
available when iterating, for example to print / export the counters.

This is useful when exporting histograms in the prometheus exposition
format that includes a running sum for histograms. These are not
included in the perf counter histograms and require a second counter.
To print both perf counters at the same time the index is needed to
match them from different collections.

Signed-off-by: Marcel Lauhoff <marcel.lauhoff@suse.com>