ci: add new sfs unit tests and fixes GC test
Adds newly added unit tests so they're found when running the
testcontainer.

It also fixes the `unittest_rgw_sfs_gc` that was throwing an exception
in certain circumstances when converting a version object.

Fix is done in this same commit as the goal is to have the newly added
tests and prevent other open PRs to be blocked.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>