ci: move tests and build script to ceph repo
This adds an extra build script that is called from a generic container
that is built from the ceph-tools repository.

The goal is to have the script that builds rgw, its tests and runs the
latest from the ceph repository itself.

It also changes the original build-radosgw.sh script from the s3gw-tools
repository to discover the sfs unit tests so we don't need to change
anything when adding a new unit test.

The build container from s3gw-tools remains the same (with a change to
run the entry point from the ceph volume) and it can be used to building
releases and for building and running unit tests in the ceph repository.

The entry point in the s3gw-tools build container uses also an
environment value that can be used to run a different script so we have
more flexibility and, for example, can run other tasks related to the
rgw that need the same environment.

All tests are now moved and executed in ceph's PR.

Fixes: https://github.com/aquarist-labs/s3gw/issues/127

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>