rgw: fix order of initialization and use in SFSGC
Fix the order of initialization and use of the SFSGC. The worker thread
can not start running until the SFSGC object has been completely
allocated, because the SFSGC object is the dout prefix provider and a
pointer to the SFSGC instance would be invalid until the constructor has
finished.
Since the worker thread might want to print something out, this would
resuult in an invalid pointer dereference.

Upon stopping, the SFSGC must query the worker thread if it has started,
because an SFSGC instance could be created without calling
`SFSGC::initialize()`. In this case calling `join()` on the worker has
to be prevented.

Fixes bugs introduced in:
f7dbf8bd3d806c8887e42975b56a01e45ccfef5a

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>