rgw/s3gw: Fix Invalid argument exception on exit with telemetry off
Shutting down the radosgw process with telemetry
disabled (--rgw-s3gw-enable-telemetry 0) caused a system error /
Invalid Argument as the telemetry class tried to stop the updater
thead that wasn't actually running. Fix this by not doing a shutdown
if we are already shut down (or never ran in the first place).

Signed-off-by: Marcel Lauhoff <marcel.lauhoff@suse.com>