rgw/sfs: avoid deleting temp db file while it's still open
When the radosgw process starts, I see the following warning:

[SQLITE] (28) file unlinked while open: [...]/build/dev/rgw.foo/s3gw.db

This happens in DBConn::check_metadata_is_compatible() because the
test_storage object still exists (i.e. database is still open)
when we call fs::remove() on the backing file.  Also, it means
we end up with a stale "s3gw.db_tmp-shm" file lying around after
the process terminates (presumably sqlite is confused).

This commit adds an extra scope so test_storage will be destroyed
before we remove the temporary database file(s).  I'm not sure if
there's a more idiomatic C++ way of doing this (it'd be nice if
there was something like python's "with" statement...)

Signed-off-by: Tim Serong <tserong@suse.com>