rgw/sfs: Conditional Copy Object
This implementens Conditional copy object.
As docs from Amazon were not clear enough on what should be the error
codes when conditions are not met I had to do a bit on investigation
and, given the error codes expected in the s3tests, plus some other
tests that I tried with other S3 vendor I conclude that all cases expect
the PRECONDITION_FAILED error, which diverts from getObject.

Fixes: https://github.com/aquarist-labs/s3gw/issues/684

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>