rgw/sfs: Fix multipart etag in response
This fixes the problem found when downloading a multipart object with
s3cmd because the `etag` was not populated in the response.

It also happened with `aws cli` tool, but in that case the tool was not
showing any error, just that the `ETag` field was not set in the
response.

When creating the final object, the `etag` should be also set in the
object's attributes, because when filling the response for `RGWGetObj`
it iterates through the attributes to fill the response.

non-multipart uploads had the `etag` already in the attributes, that's
why it was working fine in that case.

Fixes: https://github.com/aquarist-labs/s3gw/issues/658
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>