rgw/sfs: MP mark_complete: Check duplicate after update
Move the duplicate check (MP marked completed or later) untill after
the update that puts it into COMPLETE. (Ideally this would be one
SQLite statement with 'returning').

This way the transaction is a write transaction and doesn't require a
read to write lock upgrade and the default retry handler catches.

Signed-off-by: Marcel Lauhoff <marcel.lauhoff@suse.com>