rgw/sfs: Don't return version id for unversioned buckets
Avoid to return the version id when getting the object from the database
 when it wasn't explicitly required.

A few tags s3tests were failing because, for example:
```c++
int RGWDeleteObjTags::verify_permission(optional_yield y)
{
  if (!rgw::sal::Object::empty(s->object.get())) {
    auto iam_action = s->object->get_instance().empty() ?
      rgw::IAM::s3DeleteObjectTagging:
      rgw::IAM::s3DeleteObjectVersionTagging;
```
If `object->get_instance()` is not empty it changes the action to be
looked for to `DeleteObjectVersionTagging`.

The s3test is running in an unversioned bucket and setting the
`DeleteObjectTagging` action to be allowed, but it was later looking for
a different one.

Fixes: https://github.com/aquarist-labs/s3gw/issues/675
Fixes: https://github.com/aquarist-labs/s3gw/issues/695
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>